+++ src/lib/database.types.ts
@@ -2826,4 +2826,3 @@ export const Constants = {
     },
   },
 } as const
-